--- org-code-javabuilder/app/src/main/java/org/code/javabuilder/AnonymousUser.java
@@ -1,18 +1,0 @@
-package org.code.javabuilder;
-
-import java.security.Principal;
-
-/** A principal associated with a single anonymous session. */
-public class AnonymousUser implements Principal {
-
-  String name;
-
-  AnonymousUser(String name) {
-    this.name = name;
-  }
-
-  @Override
-  public String getName() {
-    return name;
-  }
-}
