--- org-code-javabuilder/app/src/main/java/org/code/javabuilder/UserProgramOutput.java
@@ -1,20 +1,0 @@
-package org.code.javabuilder;
-
-/**
- * Logical representation of the output from the user's program. Will be sent to the client as a
- * json object.
- */
-public class UserProgramOutput {
-
-  private String output;
-
-  public UserProgramOutput() {}
-
-  public UserProgramOutput(String output) {
-    this.output = output;
-  }
-
-  public String getOutput() {
-    return output;
-  }
-}
