--- org-code-javabuilder/app/src/main/resources/static/index.html
@@ -1,63 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Hello WebSocket</title>
-    <link href="/webjars/bootstrap/css/bootstrap.min.css" rel="stylesheet">
-    <link href="/main.css" rel="stylesheet">
-    <script src="/webjars/jquery/jquery.min.js"></script>
-    <script src="/webjars/sockjs-client/sockjs.min.js"></script>
-    <script src="/webjars/stomp-websocket/stomp.min.js"></script>
-    <script src="/app.js"></script>
-</head>
-<body>
-<noscript><h2 style="color: #ff0000">Seems your browser doesn't support Javascript! Websocket relies on Javascript being
-    enabled. Please enable
-    Javascript and reload this page!</h2></noscript>
-<div id="main-content" class="container">
-    <h1>Test UI for compiling and running Java with websockets</h1>
-    <div class="row">
-        <div class="col-md-6">
-            <form class="form-inline">
-                <div class="form-group">
-                    <label for="connect">WebSocket connection:</label>
-                    <button id="connect" class="btn btn-default" type="submit">Connect</button>
-                    <button id="disconnect" class="btn btn-default" type="submit" disabled="disabled">Disconnect
-                    </button>
-                </div>
-            </form>
-        </div>
-        <div class="col-md-6">
-            <form class="form-inline">
-                <div class="form-group">
-                    <label for="fileName">What is your program's file name?</label>
-                    <input type="text" id="fileName" class="form-control" placeholder="Your filename here..."></input>
-                    <label for="code">What is your code?</label>
-                    <textarea id="code" class="form-control" placeholder="Your code here..."></textarea>
-                </div>
-                <button id="send" class="btn btn-default" type="submit">Send</button>
-            </form>
-            <form class="form-inline">
-                <div class="form-group">
-                  <label for="input">Enter console input here</label>
-                  <textarea id="input" class="form-control" placeholder="> ..."></textarea>
-                </div>
-                <button id="sendInput" class="btn btn-default" type="submit">SendInput</button>
-            </form>
-        </div>
-    </div>
-    <div class="row">
-        <div class="col-md-12">
-            <table id="console" class="table table-striped">
-                <thead>
-                <tr>
-                    <th>Console Output</th>
-                </tr>
-                </thead>
-                <tbody id="output">
-                </tbody>
-            </table>
-        </div>
-    </div>
-</div>
-</body>
-</html>
