+++ org-code-javabuilder/lib/src/main/.aws-sam/temp-template.yaml
@@ -1,0 +1,9 @@
+Resources:
+  Function:
+    Type: AWS::Serverless::Function
+    Properties:
+      Timeout: 180
+      MemorySize: 256
+      Handler: org.code.javabuilder.LambdaRequestHandler
+      CodeUri: C:\Users\jmkul\IdeaProjects\java-ide\org-code-javabuilder\lib
+      Runtime: java11
