--- org-code-lambda-javabuilder/.gitignore
@@ -1,14 +1,0 @@
-# Ignore Gradle project-specific cache directory
-.gradle
-
-# Ignore Gradle build output directory
-build
-bin
-
-# Ignore Gradle auto-generated files
-app/.project
-*.classpath
-.settings
-.project
-
-.vscode/*
