--- org-code-lambda-javabuilder/dockerContainer/Dockerfile
@@ -1,5 +1,0 @@
-FROM amazoncorretto:11
-
-COPY ./*.jar /tmp/
-
-CMD java -jar /tmp/*.jar
