--- org-code-lambda-javabuilder/lib/src/main/.aws-sam/build.toml
@@ -1,11 +1,0 @@
-# This file is auto generated by SAM CLI build command
-
-[function_build_definitions]
-[function_build_definitions.d5ff8b6c-7c98-4591-84dd-67c43048db8e]
-codeuri = "C:\\Users\\jmkul\\IdeaProjects\\java-ide\\org-code-lambda-javabuilder\\lib"
-runtime = "java11"
-source_md5 = ""
-packagetype = "Zip"
-functions = ["Function"]
-
-[layer_build_definitions]
