--- org-code-lambda-javabuilder/lib/src/main/.aws-sam/temp-template.yaml
@@ -1,9 +1,0 @@
-Resources:
-  Function:
-    Type: AWS::Serverless::Function
-    Properties:
-      Timeout: 180
-      MemorySize: 512
-      Handler: org.code.lambda.javabuilder.LambdaRequestHandler
-      CodeUri: C:\Users\jmkul\IdeaProjects\java-ide\org-code-lambda-javabuilder\lib
-      Runtime: java11
